added order entity

+++ backend/src/main/java/com/eshop/backend/models/Order.java
@@ -9,10 +9,10 @@ public class Order {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private List<Product> cart;
 
-    public Order(int id, List<Product> cart) {
+    public Order(Long id, List<Product> cart) {
         this.id = id;
         this.cart = cart;
     }
@@ -21,11 +21,11 @@ public class Order {
 
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
